Refactored property search and service ticket search resolvers
