+++ graphql/schema/types/property.resolvers.ts
@@ -62,115 +62,9 @@ const property: Resolvers = {
       return [...new Set(tags)];
     },
 
-    propertiesSearch: async (_, _args, context, info) => {
-      const searchInput = {
-        searchString: _args.input.searchString.trim(),
-        options: {
-          filter: _args.input?.options?.filter ?? null,
-          facets: _args.input?.options?.facets ?? [],
-          top: _args.input?.options?.top ?? 10,
-          skip: _args.input?.options?.skip ?? 0,
-          orderBy: _args.input?.options?.orderBy ?? [],
-          hideNullResults: _args.input?.options?.hideNullResults ?? false,
-        },
-      } as PropertiesSearchInput;
-
-      const searchResults = await context.dataSources.propertySearchApi.propertiesSearch(searchInput);
-      let results = [];
-      for await (const result of searchResults?.results ?? []) {
-        results.push(result.document);
-      }
-
-      // calculate bedrooms facets
-      const bedroomsOptions = [1, 2, 3, 4, 5];
-      let bedroomsFacet = bedroomsOptions.map((option) => {
-        const found = searchResults?.facets?.bedrooms?.filter((facet) => facet.value >= option);
-        let count = 0;
-        found.forEach((f) => {
-          count += f.count;
-        });
-        return {
-          value: option + '+',
-          count: count,
-        };
-      });
-
-      // calculate bathrooms facets
-      const bathroomsOptions = [1, 1.5, 2, 3, 4, 5];
-      let bathroomsFacet = bathroomsOptions.map((option) => {
-        const found = searchResults?.facets?.bathrooms?.filter((facet) => facet.value >= option);
-        let count = 0;
-        found.forEach((f) => {
-          count += f.count;
-        });
-        return {
-          value: option + '+',
-          count: count,
-        };
-      });
-
-      // calculate update date facets
-      const periods = [7, 14, 30, 90];
-      const periodTextMaps = {
-        7: '1 week ago',
-        14: '2 weeks ago',
-        30: '1 month ago',
-        90: '3 months ago',
-      };
-
-      let periodInput = parseInt(_args.input?.options?.filter?.updatedAt); // in days
-      let updatedAtFacet = periods.map((option) => {
-        const day0 = option === periodInput ? dayjs().subtract(periodInput, 'day') : dayjs().subtract(option, 'day');
-        const found = searchResults?.facets?.updatedAt?.filter((facet) => {
-          let temp = dayjs(facet.value).diff(day0, 'day', true);
-          return temp >= 0;
-        });
-        let count = 0;
-        found.forEach((f) => {
-          count += f.count;
-        });
-        return {
-          value: periodTextMaps[option],
-          count: count,
-        };
-      });
-
-      periodInput = parseInt(_args.input?.options?.filter?.createdAt); // in days
-      let createdAtFacet = periods.map((option) => {
-        const day0 = option === periodInput ? dayjs().subtract(periodInput, 'day') : dayjs().subtract(option, 'day');
-        const found = searchResults?.facets?.createdAt?.filter((facet) => {
-          let temp = dayjs(facet.value).diff(day0, 'day', true);
-          return temp >= 0;
-        });
-        let count = 0;
-        found.forEach((f) => {
-          count += f.count;
-        });
-
-        return {
-          value: periodTextMaps[option],
-          count: count,
-        };
-      });
-
-      return {
-        propertyResults: results,
-        count: searchResults.count,
-        facets: {
-          type: searchResults.facets?.type,
-          amenities: searchResults.facets?.amenities,
-          additionalAmenitiesCategory: searchResults.facets?.['additionalAmenities/category'],
-          additionalAmenitiesAmenities: searchResults.facets?.['additionalAmenities/amenities'],
-          listedForSale: searchResults.facets?.listedForSale,
-          listedForRent: searchResults.facets?.listedForRent,
-          listedForLease: searchResults.facets?.listedForLease,
-          bedrooms: bedroomsFacet,
-          bathrooms: bathroomsFacet,
-          updatedAt: updatedAtFacet,
-          createdAt: createdAtFacet,
-          tags: searchResults.facets?.tags,
-        },
-      } as PropertySearchResult;
+    propertiesSearch: async (_, { input }, context, info) => {
+      const searchResults = await context.dataSources.propertySearchApi.propertiesSearch(input);
+      return await context.dataSources.propertySearchApi.getPropertiesSearchResults(searchResults);
     },
     getMapSasToken: async (_, _args, {dataSources}) => {
       return await dataSources.propertyMapApi.getSasToken();
